Module forest_actor::market

Structs

ActivateDealsParams

Actor
Market Actor

ClientDealProposal
ClientDealProposal is a DealProposal signed by a client

ComputeDataCommitmentParams

ComputeDataCommitmentParamsRef

ComputeDataCommitmentReturn

DealProposal
Note: Deal Collateral is only released and returned to clients and miners when the storage deal stops counting towards power. In the current iteration, it will be released when the sector containing the storage deals expires, even though some storage deals can expire earlier than the sector does. Collaterals are denominated in PerEpoch to incur a cost for self dealing or minimal deals that last for a long time. Note: ClientCollateralPerEpoch may not be needed and removed pending future confirmation. There will be a Minimum value for both client and provider deal collateral.

DealState

OnMinerSectorsTerminateParams

OnMinerSectorsTerminateParamsRef

PublishStorageDealsParams

PublishStorageDealsReturn

SectorDataSpec

SectorDeals

SectorWeights

State
Market actor state

VerifyDealsForActivationParams

VerifyDealsForActivationParamsRef

VerifyDealsForActivationReturn

WithdrawBalanceParams

Enums

Method
Market actor methods available

Constants

PIECE_CID_PREFIX
Cid prefix for piece Cids

PROPOSALS_AMT_BITWIDTH

STATES_AMT_BITWIDTH

Functions

validate_and_compute_deal_weight

validate_deals_for_activation
Validates a collection of deal dealProposals for activation, and returns their combined weight, split into regular deal weight and verified deal weight.

Type Definitions

DealArray
A specialization of a array to deals.

DealMetaArray
A specialization of a array to deals.