Function forest_actor::market::validate_deals_for_activation

pub fn validate_deals_for_activation<BS>(
    st: &State, 
    store: &BS, 
    deal_ids: &[DealID], 
    miner_addr: &Address, 
    sector_expiry: ChainEpoch, 
    curr_epoch: ChainEpoch
) -> Result<(BigInt, BigInt, u64), Box<dyn StdError>> where
    BS: BlockStore, 

Validates a collection of deal dealProposals for activation, and returns their combined weight, split into regular deal weight and verified deal weight.