Struct forest_actor::market::DealProposal

pub struct DealProposal {
    pub piece_cid: Cid,
    pub piece_size: PaddedPieceSize,
    pub verified_deal: bool,
    pub client: Address,
    pub provider: Address,
    pub label: String,
    pub start_epoch: ChainEpoch,
    pub end_epoch: ChainEpoch,
    pub storage_price_per_epoch: TokenAmount,
    pub provider_collateral: TokenAmount,
    pub client_collateral: TokenAmount,
}

Note: Deal Collateral is only released and returned to clients and miners when the storage deal stops counting towards power. In the current iteration, it will be released when the sector containing the storage deals expires, even though some storage deals can expire earlier than the sector does. Collaterals are denominated in PerEpoch to incur a cost for self dealing or minimal deals that last for a long time. Note: ClientCollateralPerEpoch may not be needed and removed pending future confirmation. There will be a Minimum value for both client and provider deal collateral.

Fields
piece_cid: Cidpiece_size: PaddedPieceSizeverified_deal: boolclient: Addressprovider: Addresslabel: String

Arbitrary client chosen label to apply to the deal

start_epoch: ChainEpochend_epoch: ChainEpochstorage_price_per_epoch: TokenAmountprovider_collateral: TokenAmountclient_collateral: TokenAmount
Implementations

impl DealProposal

pub fn duration(&self) -> ChainEpoch

pub fn weight(&self) -> DealWeight

Computes weight for a deal proposal, which is a function of its size and duration.

pub fn total_storage_fee(&self) -> TokenAmount

pub fn client_balance_requirement(&self) -> TokenAmount

pub fn provider_balance_requirement(&self) -> &TokenAmount