Function forest_actor::miner::initial_pledge_for_power

pub fn initial_pledge_for_power(
    qa_power: &StoragePower, 
    baseline_power: &StoragePower, 
    reward_estimate: &FilterEstimate, 
    network_qa_power_estimate: &FilterEstimate, 
    circulating_supply: &TokenAmount
) -> TokenAmount

Computes the pledge requirement for committing new quality-adjusted power to the network, given the current network total and baseline power, per-epoch reward, and circulating token supply. The pledge comprises two parts:

storage pledge, aka IP base: a multiple of the reward expected to be earned by newly-committed power

consensus pledge, aka additional IP: a pro-rata fraction of the circulating money supply

IP = IPBase(t) + AdditionalIP(t) IPBase(t) = BR(t, InitialPledgeProjectionPeriod) AdditionalIP(t) = LockTarget(t)*PledgeShare(t) LockTarget = (LockTargetFactorNum / LockTargetFactorDenom) * FILCirculatingSupply(t) PledgeShare(t) = sectorQAPower / max(BaselinePower(t), NetworkQAPower(t))