Struct forest_actor::miner::ExpirationSet

pub struct ExpirationSet {
    pub on_time_sectors: BitField,
    pub early_sectors: BitField,
    pub on_time_pledge: TokenAmount,
    pub active_power: PowerPair,
    pub faulty_power: PowerPair,
}

ExpirationSet is a collection of sector numbers that are expiring, either due to expected “on-time” expiration at the end of their life, or unexpected “early” termination due to being faulty for too long consecutively. Note that there is not a direct correspondence between on-time sectors and active power; a sector may be faulty but expiring on-time if it faults just prior to expected termination. Early sectors are always faulty, and active power always represents on-time sectors.

Fields
on_time_sectors: BitField

Sectors expiring “on time” at the end of their committed life

early_sectors: BitField

Sectors expiring “early” due to being faulty for too long

on_time_pledge: TokenAmount

Pledge total for the on-time sectors

active_power: PowerPair

Power that is currently active (not faulty)

faulty_power: PowerPair

Power that is currently faulty

Implementations

impl ExpirationSet

pub fn empty() -> Self

pub fn add(
&mut self,
on_time_sectors: &BitField,
early_sectors: &BitField,
on_time_pledge: &TokenAmount,
active_power: &PowerPair,
faulty_power: &PowerPair
)

Adds sectors and power to the expiration set in place.

pub fn remove(
&mut self,
on_time_sectors: &BitField,
early_sectors: &BitField,
on_time_pledge: &TokenAmount,
active_power: &PowerPair,
faulty_power: &PowerPair
) -> Result<(), Box<dyn StdError>>

Removes sectors and power from the expiration set in place.

pub fn is_empty(&self) -> bool

A set is empty if it has no sectors. The power and pledge are not checked, but are expected to be zero.

pub fn len(&self) -> usize

Counts all sectors in the expiration set.

pub fn validate_state(&self) -> Result<(), &'static str>

validates a set of assertions that must hold for expiration sets