Module forest_actor::multisig

Structs

Actor
Multisig Actor

AddSignerParams
Add signer params.

ApproveReturn
Parameters for approve and cancel multisig functions.

ChangeNumApprovalsThresholdParams
Propose method call parameters

ConstructorParams
Constructor parameters for multisig actor.

LockBalanceParams
Lock balance call params.

ProposalHashData
Data for a BLAKE2B-256 to be attached to methods referencing proposals via TXIDs. Ensures the existence of a cryptographic reference to the original proposal. Useful for offline signers and for protection when reorgs change a multisig TXID.

ProposeParams
Propose method call parameters.

ProposeReturn
Propose method call return.

RemoveSignerParams
Remove signer params.

State
Multisig actor state

SwapSignerParams
Swap signer multisig method params

Transaction
Transaction type used in multisig actor

TxnID
Transaction ID type

TxnIDParams
Parameters for approve and cancel multisig functions.

Enums

Method
Multisig actor methods available