Struct forest_actor::util::BalanceTable

pub struct BalanceTable<'a, BS>(_);

Balance table which handles getting and updating token balances specifically

Implementations

impl<'a, BS> BalanceTable<'a, BS> where
BS: BlockStore,

pub fn new(bs: &'a BS) -> Self

Initializes a new empty balance table

pub fn from_root(bs: &'a BS, cid: &Cid) -> Result<Self, Error>

Initializes a balance table from a root Cid

pub fn root(&mut self) -> Result<Cid, Error>

Retrieve root from balance table

pub fn get(&self, key: &Address) -> Result<TokenAmount, Box<dyn StdError>>

Gets token amount for given address in balance table

pub fn add(
&mut self,
key: &Address,
value: &TokenAmount
) -> Result<(), Box<dyn StdError>>

Adds token amount to previously initialized account.

pub fn subtract_with_minimum(
&mut self,
key: &Address,
req: &TokenAmount,
floor: &TokenAmount
) -> Result<TokenAmount, Box<dyn StdError>>

Subtracts up to the specified amount from a balance, without reducing the balance below some minimum. Returns the amount subtracted (always positive or zero).

pub fn must_subtract(
&mut self,
key: &Address,
req: &TokenAmount
) -> Result<(), Box<dyn StdError>>

Subtracts value from a balance, and errors if full amount was not substracted.

pub fn total(&self) -> Result<TokenAmount, Box<dyn StdError>>

Returns total balance held by this balance table